requirements.txt
pydantic
loguru
chardet
unidecode
colorama
rich
pandas
openpyxl
xlwt
python-docx
pdfminer.six
pillow
pytesseract
requests
send2trash
pyyaml
